Fix tag size

+++ ssf/src/ir/constructor.rs
@@ -3,11 +3,11 @@ use crate::types::{self, Algebraic, Type};
 #[derive(Clone, Debug, PartialEq)]
 pub struct Constructor {
     type_: Algebraic,
-    index: usize,
+    index: u64,
 }
 
 impl Constructor {
-    pub fn new(type_: Algebraic, index: usize) -> Self {
+    pub fn new(type_: Algebraic, index: u64) -> Self {
         Self { type_, index }
     }
 
@@ -19,7 +19,7 @@ impl Constructor {
         &self.type_.constructors()[&self.index]
     }
 
-    pub fn index(&self) -> usize {
+    pub fn index(&self) -> u64 {
         self.index
     }
